refactor(db_sync): split monolithic db_sync.py into focused modules
- Split db_sync.py into postgres.py, artifacts.py, charts.py, and snapshots.py for better separation of concerns.
- Updated aws_db.py to use DbSyncConfig instead of global variables. All functions re-exported from db_sync.py
for backward compatibility.
